feat: support new ship layout format

+++ ShipEditor/Scripts/Model/Extensions.cs
@@ -2,7 +2,7 @@
 {
 	public static class CellIndex
 	{
-		public static ulong FromXY(int x, int y) => ((ulong)y << 32) + (uint)x;
+		public static ulong FromXY(int x, int y) => ((ulong)(uint)y << 32) + (uint)x;
 		
 		public static void GetXY(ulong index, out int x, out int y)
 		{
